RF: Generate derivative cache per-BOLD, get one-way transforms
