+++ fmriprep/utils/bids.py
@@ -31,6 +31,7 @@ from collections import defaultdict
 from pathlib import Path
 
 from bids.layout import BIDSLayout
+from bids.utils import listify
 from packaging.version import Version
 
 from ..data import load
@@ -38,7 +39,8 @@ from ..data import load
 
 def collect_derivatives(
     derivatives_dir: Path,
-    subject_id: str,
+    entities: dict,
+    fieldmap_id: str | None,
     spec: dict | None = None,
     patterns: ty.List[str] | None = None,
 ):
@@ -59,23 +61,20 @@ def collect_derivatives(
 
     # search for both boldrefs
     for k, q in spec["baseline"].items():
-        q["subject"] = subject_id
-        item = layout.get(return_type='filename', **q)
+        query = {**q, **entities}
+        item = layout.get(return_type='filename', **query)
         if not item:
             continue
         derivs_cache["%s_boldref" % k] = item[0] if len(item) == 1 else item
 
-    for xfm in spec['transforms']:
-        for k, q in spec['transforms'][xfm].items():
-            # flip the X2Y naming for opposite direction
-            if k == 'reverse':
-                xfm = '2'.join(xfm.split('2')[::-1])
-            q = q.copy()
-            q['subject'] = subject_id
-            item = layout.get(return_type='filename', **q)
-            if not item:
-                continue
-            derivs_cache[xfm] = item[0] if len(item) == 1 else item
+    for xfm, q in spec['transforms'].items():
+        query = {**q, **entities}
+        if xfm == "boldref2fmap":
+            query["to"] = fieldmap_id
+        item = layout.get(return_type='filename', **q)
+        if not item:
+            continue
+        derivs_cache[xfm] = item[0] if len(item) == 1 else item
     return derivs_cache
 
 
@@ -291,3 +290,37 @@ def check_pipeline_version(pipeline_name, cvers, data_desc):
         dvers = desc.get("PipelineDescription", {}).get("Version", "0+unknown")
     if Version(cvers).public != Version(dvers).public:
         return "Previous output generated by version {} found.".format(dvers)
+
+
+def extract_entities(file_list):
+    """
+    Return a dictionary of common entities given a list of files.
+
+    Examples
+    --------
+    >>> extract_entities("sub-01/anat/sub-01_T1w.nii.gz")
+    {'subject': '01', 'suffix': 'T1w', 'datatype': 'anat', 'extension': '.nii.gz'}
+    >>> extract_entities(["sub-01/anat/sub-01_T1w.nii.gz"] * 2)
+    {'subject': '01', 'suffix': 'T1w', 'datatype': 'anat', 'extension': '.nii.gz'}
+    >>> extract_entities(["sub-01/anat/sub-01_run-1_T1w.nii.gz",
+    ...                   "sub-01/anat/sub-01_run-2_T1w.nii.gz"])
+    {'subject': '01', 'run': [1, 2], 'suffix': 'T1w', 'datatype': 'anat', 'extension': '.nii.gz'}
+
+    """
+    from collections import defaultdict
+
+    from bids.layout import parse_file_entities
+
+    entities = defaultdict(list)
+    for e, v in [
+        ev_pair for f in listify(file_list) for ev_pair in parse_file_entities(f).items()
+    ]:
+        entities[e].append(v)
+
+    def _unique(inlist):
+        inlist = sorted(set(inlist))
+        if len(inlist) == 1:
+            return inlist[0]
+        return inlist
+
+    return {k: _unique(v) for k, v in entities.items()}
