+++ fmriprep/workflows/bold/fit.py
@@ -44,6 +44,7 @@ from sdcflows.workflows.apply.correction import init_unwarp_wf
 from sdcflows.workflows.apply.registration import init_coeff2epi_wf
 
 from ... import config
+from ...utils.bids import extract_entities
 
 # BOLD workflows
 from .hmc import init_bold_hmc_wf
@@ -459,40 +460,6 @@ def _to_join(in_file, join_file):
         return in_file
     res = JoinTSVColumns(in_file=in_file, join_file=join_file).run()
     return res.outputs.out_file
-
-
-def extract_entities(file_list):
-    """
-    Return a dictionary of common entities given a list of files.
-
-    Examples
-    --------
-    >>> extract_entities("sub-01/anat/sub-01_T1w.nii.gz")
-    {'subject': '01', 'suffix': 'T1w', 'datatype': 'anat', 'extension': '.nii.gz'}
-    >>> extract_entities(["sub-01/anat/sub-01_T1w.nii.gz"] * 2)
-    {'subject': '01', 'suffix': 'T1w', 'datatype': 'anat', 'extension': '.nii.gz'}
-    >>> extract_entities(["sub-01/anat/sub-01_run-1_T1w.nii.gz",
-    ...                   "sub-01/anat/sub-01_run-2_T1w.nii.gz"])
-    {'subject': '01', 'run': [1, 2], 'suffix': 'T1w', 'datatype': 'anat', 'extension': '.nii.gz'}
-
-    """
-    from collections import defaultdict
-
-    from bids.layout import parse_file_entities
-
-    entities = defaultdict(list)
-    for e, v in [
-        ev_pair for f in listify(file_list) for ev_pair in parse_file_entities(f).items()
-    ]:
-        entities[e].append(v)
-
-    def _unique(inlist):
-        inlist = sorted(set(inlist))
-        if len(inlist) == 1:
-            return inlist[0]
-        return inlist
-
-    return {k: _unique(v) for k, v in entities.items()}
 
 
 def _select_ref(sbref_files, boldref_files):
